more code cleanup

+++ pet-race-job/pet_race_job/model/race_normal.py
@@ -24,7 +24,7 @@ class RaceNormal(Model):
     petCategoryName = columns.Text(required=False)
     petCategoryId = columns.UUID(primary_key=True, default=uuid.uuid4)
     currentTime = columns.DateTime()
-    normals = columns.List(value_type=columns.Float())
+    normals = columns.List(value_type=columns.Decimal())
     normalLoc = columns.Float()
     normalScale = columns.Float()
     normalSize = columns.Integer()
